+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.1.0
+## ADDED
+- ParagraphToolbar (Bold, italic, Underline, HTML Code view)
+- Styled toolbar, grab bar, started general settings bar
+- Delete button on general settings bar
+- MiniModal
+
 # 0.0.1
 ## ADDED
 - Started basic Block